--- src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/applepay/applePayExpress.js
@@ -1,3 +1,0 @@
-const applePayExpressModule = require('./common');
-
-applePayExpressModule.init();
