+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/applePayExpress.js
@@ -1,0 +1,3 @@
+const applePayExpressModule = require('./applePayExpressCommon');
+
+applePayExpressModule.init();
